requirements.txt
# === Orchestrator core (M1 minimum) ===
anthropic>=0.36.0
supabase>=2.8.0
pydantic>=2.8.0
pyyaml>=6.0.2
fastapi>=0.114.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
loguru>=0.7.0
tenacity>=9.0.0            # retries with backoff

# === Research (M1) ===
serpapi>=0.1.5
playwright>=1.46.0
beautifulsoup4>=4.12.0
youtube-transcript-api>=0.6.2

# === Telegram approvals (M1) ===
python-telegram-bot>=21.5

# === Infra (M2+) ===
celery>=5.4.0
redis>=5.1.0
boto3>=1.35.0              # MinIO / S3
minio>=7.2.0
google-api-python-client>=2.143.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# === Media (M2+) — installed on render nodes, not the orchestrator ===
# chatterbox-tts, kokoro, f5-tts, TTS, diffusers, moviepy, ffmpeg-python,
# scenedetect, audiocraft, demucs, pedalboard, openai-whisper, realesrgan
# See docs/render-nodes.md for per-node install instructions.

# === Dev ===
pytest>=8.3.0
pytest-asyncio>=0.24.0
ruff>=0.7.0
mypy>=1.11.0
